Add a better __repr__ to code_objects for debugging
Instead of simply showing the argument of each opcode(the index
in the corresponding list at which the object lies), print out
the actual object at the index of each opcode's instruction (or
"[no args]" if the instruction has no arguments).

Also, show the position of each opcode in their respective lists
to make debugging jump instructions easier.