Change imports to be realtive. Add context.py to tests/
To correctly run Schemey, the imports
had to be changed to realtive. This allowed Python
to correctly find and execute each imported module.

Also, to correctly import source files from inside of the
tests/ package, a helper module - context.py - was created.
This helper module appends the path of the src/ package to
sys.path, allowing python to find it.